🚧Implement various improvements (#262)
* Start work for issue #144

* ci: changed any description text reference `twitter` to `x`

* docs: update all twitter urls and text to x

* ci!: change twitter related secret names to x

* refactor: rename twitter env vars for send release tweet script

* ide: update twitter env vars for launch config

* chore: use the release notes dir path workflow input

Update the release notes exit step and associated script to make use of the
relative release notes dir path.  It was not being used before at all.

* refactor: change code to meet coding standards

* refactor: update name of func name var

* ci: remove step with unused output from release workflow

* ci: fix github messages in workflow step

* config: configure git and vscode editor settings for line endings

configures git attributes for consistent line endings and updates
vscode settings for editor preferences, deno, and cspell words

* refactor: change code to meet coding standards

* cleanup: remove unused modules and types

* refactor: improve the isNothing function

* refactor: rename guard class to paramguards and use functions

renames the guard class to paramguards to better reflect its
purpose and updates its functions to standalone functions. this
improves code clarity and simplifies usage across the codebase.

* refactor: refactor core utilities and scripts

refactors core utilities to remove `utils` class and update
scripts to use standalone functions for better modularity
and readability

* cleanup: remove issue template fields and labels

removes unnecessary issue template fields and associated labels to
streamline issue creation and improve clarity

* ci: change name of send release tweet workflow input

* ci: rename send release tweet script

* ci: rename release tweet environment variables

* ide: rename release tweet env vars in launch config

* ci: refactor release tweet builder name

* ci: rename release tweet builder file

* config: rename release tweet post template file

* cicd: update variable names and messages to use "x" instead of "tweet"

updates variable names and user messages in the script to reflect
the use of "x" for broadcasting instead of the outdated "tweet". this
clarifies the functionality and aligns it with the updated terminology.